Create PowerOfTwo.js
Given a positive integer n, which is equal to 2, raised to some power k:

powerof2

Find this degree (number k).

Write a function named powerOf2 that takes n as an argument and returns k such that 2 to the k power is n. When solving, use a while loop and operations of division or multiplication. It is forbidden to use methods of the Math object.