Create Reverse_array_ofdigits.js
Coding Practice
An array of digits of a positive number has to be written in reverse order.

A positive integer is given. Get an array of its digits, written in reverse order (from the last digit to the first).

Write a function called arrayOfDigits that takes the number n as an argument and returns an array of its digits in reverse order. It is forbidden to use the split, reverse methods. Use a do while loop in your solution.